requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==5.0
django-ckeditor==6.7.0
django-js-asset==2.2.0
gunicorn==21.2.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
wheel==0.41.2
whitenoise==6.6.0
